working on update mutation
