--- server/utils/validation.js
@@ -1,15 +1,0 @@
-
-import Client from "../models/clientModel.js"
-
-export const findExistingClient = async (name,email,phone) => {
-    
-    const existingClient = await Client.findOne({
-        $or: [
-            { name },
-            { email },
-            { phone }
-        ]
-    });
-    
-    return existingClient;
-}
